# CUID generator optimization