Optimisation of CUID generation via the use of a buffer containing random data. This allows for fewer calls to the system to obtain entropy.

### Metric
1,000,000 iterations

Before:
 * UUID: ~550 ms
 * CUID: ~1200 ms

After:
 * UUID: ~550 ms
 * CUID: ~420 ms

**Machine specs :** _AMD Ryzen 7 3700X 8-Core 3.60 Ghz /  16 GB Ram / Windows 11_